Refactor galaxy.web.form_builder into galaxy.util.form_builder.
We have other utilities for web content in galaxy.util like sanitize_html, so while I'm not eager to put web stuff in there it is not bringing any dependencies into galaxy's util package. The upshot here is that now model doesn't need to have dependencies on galaxy.web (other than implicitly through galaxy.managers.tags - but that really has to be refactored either way).

Contains some more cleanup related to url_for (see #7521 for context).
